backend/requirements.txt
fastapi
pydantic
duckdb
pymongo
passlib[bcrypt]
dependency-injector
uvicorn
jose
python-dotenv
pydantic-settings